requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data Processing
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Data Visualization & Analytics
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Machine Learning
scikit-learn==1.3.2
scipy==1.11.4

# AWS Services
boto3==1.34.0
botocore==1.34.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# API Documentation
python-multipart==0.0.6
